fix file sizing
