+++ modal_training_gym/cli/run.py
@@ -266,7 +266,7 @@ def _download_trace_step(
             "file_name": file_name,
             "samples": rollout.total,
             "mean_reward": rollout.mean,
-            "size_bytes": len(data),
+            "export_size_bytes": len(data),
         },
         len(data),
     )
@@ -278,7 +278,7 @@ def download_run_traces(
     out: str,
     step: str | None,
     dry_run: bool,
-    yes: bool,
+    skip_confirmation: bool,
     json_output: bool,
 ) -> None:
     """Download selected rollout payloads and write a local trace manifest."""
@@ -327,7 +327,7 @@ def download_run_traces(
                 summaries_by_step[value] for value in sorted(selected_steps)
             ]
 
-        known_sizes = [summary.size_bytes for summary in selected_summaries]
+        known_sizes = [summary.export_size_bytes for summary in selected_summaries]
         total_size = (
             sum(size for size in known_sizes if size is not None)
             if all(size is not None for size in known_sizes)
@@ -339,19 +339,19 @@ def download_run_traces(
             "dry_run": dry_run,
             "step_count": len(selected_summaries),
             "sample_count": sum(summary.total for summary in selected_summaries),
-            "size_bytes": total_size,
+            "export_size_bytes": total_size,
             "steps": [
                 {
                     "step": summary.rollout_id,
                     "file_name": f"step_{summary.rollout_id:04d}.json",
                     "samples": summary.total,
                     "mean_reward": summary.mean,
-                    "size_bytes": summary.size_bytes,
+                    "export_size_bytes": summary.export_size_bytes,
                 }
                 for summary in selected_summaries
             ],
         }
-        estimated_size = report["size_bytes"]
+        estimated_size = report["export_size_bytes"]
         if not isinstance(estimated_size, int):
             estimated_size_text = "unknown size"
         else:
@@ -367,7 +367,7 @@ def download_run_traces(
                 )
             return
 
-        if not yes:
+        if not skip_confirmation:
             confirm_or_require_yes(
                 f"Download {report['step_count']} steps "
                 f"({report['sample_count']} samples, "
@@ -418,7 +418,7 @@ def download_run_traces(
         finally:
             shutil.rmtree(staging_path, ignore_errors=True)
 
-    report["size_bytes"] = downloaded_size
+    report["export_size_bytes"] = downloaded_size
     report["steps"] = downloaded_steps
     if json_output:
         print_json(report)
@@ -1049,7 +1049,7 @@ def trace_command(
         out=out,
         step=step,
         dry_run=dry_run,
-        yes=yes,
+        skip_confirmation=yes,
         json_output=json_output,
     )
